requirements.txt
streamlit
spacy
pymupdf
python-docx
pandas
